unit test moved to pool manager
