+++ test/test_connection.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import datetime
 import socket
-import sys
 import typing
 from http.client import ResponseNotReady
 from unittest import mock
@@ -236,13 +235,3 @@ class TestConnection:
         # Should error if a request has not been sent
         with pytest.raises(ResponseNotReady):
             conn.getresponse()
-
-    def test_http_client_connect_audit_event(self) -> None:
-        if hasattr(sys, "addaudithook"):
-
-            def _hook(event: str, args: tuple) -> None:  # type: ignore[type-arg]
-                if event == "http.client.connect":
-                    assert event == "http.client.connect"
-
-            sys.addaudithook(_hook)
-            HTTPConnection("google.com", port=80)
